+++ common/src/main/java/dev/ripio/cobbleloots/data/custom/filter/CobblelootsSourceFilter.java
@@ -2,7 +2,6 @@ package dev.ripio.cobbleloots.data.custom.filter;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
-import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.levelgen.structure.Structure;
 import net.minecraft.world.level.material.Fluid;
 
@@ -11,7 +10,7 @@ import java.util.List;
 public class CobblelootsSourceFilter {
   private final int weight;
   private final TagKey<Structure> structure;
-  private final TagKey<Biome> biome;
+  private final CobbleloootsBiomeFilter biome;
   private final List<ResourceLocation> dimension;
   private final CobblelootsBlockFilter block;
   private final TagKey<Fluid> fluid;
@@ -22,7 +21,7 @@ public class CobblelootsSourceFilter {
   private final CobblelootsDateFilter date;
   private final CobblelootsPokeRodFilter pokeRod;
 
-  public CobblelootsSourceFilter(int weight, TagKey<Structure> structure, TagKey<Biome> biome,
+  public CobblelootsSourceFilter(int weight, TagKey<Structure> structure, CobbleloootsBiomeFilter biome,
       List<ResourceLocation> dimension, CobblelootsBlockFilter block, TagKey<Fluid> fluid,
       CobblelootsPositionFilter position, CobblelootsLightFilter light, CobblelootsTimeFilter time,
       CobblelootsWeatherFilter weather, CobblelootsDateFilter date, CobblelootsPokeRodFilter pokeRod) {
@@ -48,7 +47,7 @@ public class CobblelootsSourceFilter {
     return structure;
   }
 
-  public TagKey<Biome> getBiome() {
+  public CobbleloootsBiomeFilter getBiome() {
     return biome;
   }